+++ SQLComponents/SQLVariantOperatorSub.cpp
@@ -56,7 +56,7 @@ static SQL_OperUShortSubChar(SQLVariant& p_left,SQLVariant& p_right)
 SQLVariant 
 static SQL_OperSLongSubChar(SQLVariant& p_left,SQLVariant& p_right)
 {
-  long result = p_left.GetAsSLong() - p_right.GetAsSLong();
+  int result = p_left.GetAsSLong() - p_right.GetAsSLong();
   return SQLVariant(result);
 }
 SQLVariant 
@@ -162,7 +162,7 @@ static SQL_OperUShortSubSShort(SQLVariant& p_left,SQLVariant& p_right)
 
 SQLVariant static SQL_OperSLongSubSShort(SQLVariant& p_left,SQLVariant& p_right)
 {
-  long result = p_left.GetAsSLong();
+  int result = p_left.GetAsSLong();
   result -= p_right.GetAsSLong();
   return SQLVariant(result);
 }
@@ -271,7 +271,7 @@ static SQL_OperUShortSubUShort(SQLVariant& p_left,SQLVariant& p_right)
 SQLVariant 
 static SQL_OperSLongSubUShort(SQLVariant& p_left,SQLVariant& p_right)
 {
-  long result = p_left.GetAsSLong();
+  int result = p_left.GetAsSLong();
   result -= p_right.GetAsSLong();
   return SQLVariant(result);
 }
@@ -356,7 +356,7 @@ static SQL_OperNumSubUShort(SQLVariant& p_left,SQLVariant& p_right)
 SQLVariant 
 static SQL_OperCharSubSLong(SQLVariant& p_left,SQLVariant& p_right)
 {
-  long result = p_left.GetAsSLong();
+  int result = p_left.GetAsSLong();
   result -= p_right.GetAsSLong();
   return SQLVariant(result);
 }
@@ -380,7 +380,7 @@ static SQL_OperUShortSubSLong(SQLVariant& p_left,SQLVariant& p_right)
 SQLVariant 
 static SQL_OperSLongSubSLong(SQLVariant& p_left,SQLVariant& p_right)
 {
-  long result = p_left.GetAsSLong();
+  int result = p_left.GetAsSLong();
   result -= p_right.GetAsSLong();
   return SQLVariant(result);
 }
@@ -598,7 +598,7 @@ static SQL_OperUShortSubFloat(SQLVariant& p_left,SQLVariant& p_right)
 SQLVariant 
 static SQL_OperSLongSubFloat(SQLVariant& p_left,SQLVariant& p_right)
 {
-  long result = p_left.GetAsSLong();
+  int result = p_left.GetAsSLong();
   result -= p_right.GetAsSLong();
   return SQLVariant(result);
 }
@@ -707,7 +707,7 @@ static SQL_OperUShortSubDouble(SQLVariant& p_left,SQLVariant& p_right)
 SQLVariant 
 static SQL_OperSLongSubDouble(SQLVariant& p_left,SQLVariant& p_right)
 {
-  long result = p_left.GetAsSLong();
+  int result = p_left.GetAsSLong();
   result -= p_right.GetAsSLong();
   return SQLVariant(result);
 }
@@ -825,7 +825,7 @@ static SQL_OperUShortSubBit(SQLVariant& p_left,SQLVariant& p_right)
 SQLVariant 
 static SQL_OperSLongSubBit(SQLVariant& p_left,SQLVariant& p_right)
 {
-  long result = p_left.GetAsSLong();
+  int result = p_left.GetAsSLong();
   if(p_right.GetAsBit())
   {
     --result;
@@ -958,7 +958,7 @@ static SQL_OperUShortSubSTiny(SQLVariant& p_left,SQLVariant& p_right)
 SQLVariant 
 static SQL_OperSLongSubSTiny(SQLVariant& p_left,SQLVariant& p_right)
 {
-  long result = p_left.GetAsSLong();
+  int result = p_left.GetAsSLong();
   result -= p_right.GetAsSLong();
   return SQLVariant(result);
 }
@@ -1067,7 +1067,7 @@ static SQL_OperUShortSubUTiny(SQLVariant& p_left,SQLVariant& p_right)
 SQLVariant 
 static SQL_OperSLongSubUTiny(SQLVariant& p_left,SQLVariant& p_right)
 {
-  long result = p_left.GetAsSLong();
+  int result = p_left.GetAsSLong();
   result -= p_right.GetAsSLong();
   return SQLVariant(result);
 }
@@ -1176,7 +1176,7 @@ static SQL_OperUShortSubSBig(SQLVariant& p_left,SQLVariant& p_right)
 SQLVariant 
 static SQL_OperSLongSubSBig(SQLVariant& p_left,SQLVariant& p_right)
 {
-  long result = p_left.GetAsSLong();
+  int result = p_left.GetAsSLong();
   result -= p_right.GetAsSLong();
   return SQLVariant(result);
 }
@@ -1285,7 +1285,7 @@ static SQL_OperUShortSubUBig(SQLVariant& p_left,SQLVariant& p_right)
 SQLVariant 
 static SQL_OperSLongSubUBig(SQLVariant& p_left,SQLVariant& p_right)
 {
-  long result = p_left.GetAsSLong();
+  int result = p_left.GetAsSLong();
   result -= p_right.GetAsSLong();
   return SQLVariant(result);
 }
@@ -1396,7 +1396,7 @@ static SQL_OperUShortSubNum(SQLVariant& p_left,SQLVariant& p_right)
 SQLVariant 
 static SQL_OperSLongSubNum(SQLVariant& p_left,SQLVariant& p_right)
 {
-  long num = p_left.GetAsSLong() - p_right.GetAsBCD().AsLong();
+  int num = p_left.GetAsSLong() - p_right.GetAsBCD().AsLong();
   SQLVariant var(num);
   return var;
 }
