+++ UnitTest/CalculateTest.cpp
@@ -56,7 +56,7 @@ namespace OperatorUnitTest
       variants[CT_CHAR]     = SQLVariant("42");
       variants[CT_SSHORT]   = SQLVariant((short)42);
       variants[CT_USHORT]   = SQLVariant((unsigned short)42);
-      variants[CT_SLONG]    = SQLVariant((long)42);
+      variants[CT_SLONG]    = SQLVariant(42);
       variants[CT_ULONG]    = SQLVariant((unsigned int)42);
       variants[CT_FLOAT]    = SQLVariant((float)42.0);
       variants[CT_DOUBLE]   = SQLVariant((double)42.0);
@@ -70,7 +70,7 @@ namespace OperatorUnitTest
       other[CT_CHAR]     = SQLVariant("5");
       other[CT_SSHORT]   = SQLVariant((short)5);
       other[CT_USHORT]   = SQLVariant((unsigned short)5);
-      other[CT_SLONG]    = SQLVariant((long)5);
+      other[CT_SLONG]    = SQLVariant(5);
       other[CT_ULONG]    = SQLVariant((unsigned int)5);
       other[CT_FLOAT]    = SQLVariant((float)5);
       other[CT_DOUBLE]   = SQLVariant((double)5);
